+++ monkey/common/cloud/instance.py
@@ -1,6 +1,3 @@
-from common.cloud.environment_names import Environment
-
-
 class CloudInstance(object):
     """
     This is an abstract class which represents a cloud instance.
@@ -9,7 +6,4 @@ class CloudInstance(object):
     """
 
     def is_instance(self) -> bool:
-        raise NotImplementedError()
-
-    def get_cloud_provider_name(self) -> Environment:
         raise NotImplementedError()