+++ monkey/monkey_island/cc/models/monkey.py
@@ -14,7 +14,6 @@ from mongoengine import (
     StringField,
 )
 
-from common.cloud import environment_names
 from monkey_island.cc.models.command_control_channel import CommandControlChannel
 from monkey_island.cc.models.monkey_ttl import MonkeyTtl, create_monkey_ttl_document
 from monkey_island.cc.server_utils.consts import DEFAULT_MONKEY_TTL_EXPIRY_DURATION_IN_SECONDS
@@ -55,14 +54,8 @@ class Monkey(Document):
     tunnel = ReferenceField("self")
     command_control_channel = EmbeddedDocumentField(CommandControlChannel)
 
-    # Environment related fields
-    environment = StringField(
-        default=environment_names.Environment.UNKNOWN.value,
-        choices=environment_names.ALL_ENVIRONMENTS_NAMES,
-    )
-    aws_instance_id = StringField(
-        required=False
-    )  # This field only exists when the monkey is running on an AWS
+    # This field only exists when the monkey is running on an AWS
+    aws_instance_id = StringField(required=False)
 
     # instance. See https://github.com/guardicore/monkey/issues/426.