+++ monkey/monkey_island/cc/services/config_schema/monkey.py
@@ -1,6 +1,5 @@
 from common.common_consts.system_info_collectors_names import (
     AWS_COLLECTOR,
-    ENVIRONMENT_COLLECTOR,
     HOSTNAME_COLLECTOR,
     MIMIKATZ_COLLECTOR,
     PROCESS_LIST_COLLECTOR,
@@ -88,7 +87,6 @@ MONKEY = {
                     "uniqueItems": True,
                     "items": {"$ref": "#/definitions/system_info_collector_classes"},
                     "default": [
-                        ENVIRONMENT_COLLECTOR,
                         AWS_COLLECTOR,
                         HOSTNAME_COLLECTOR,
                         PROCESS_LIST_COLLECTOR,