+++ monkey/monkey_island/cc/services/telemetry/processing/system_info_collectors/system_info_telemetry_dispatcher.py
@@ -3,15 +3,11 @@ import typing
 
 from common.common_consts.system_info_collectors_names import (
     AWS_COLLECTOR,
-    ENVIRONMENT_COLLECTOR,
     HOSTNAME_COLLECTOR,
     PROCESS_LIST_COLLECTOR,
 )
 from monkey_island.cc.services.telemetry.processing.system_info_collectors.aws import (
     process_aws_telemetry,
-)
-from monkey_island.cc.services.telemetry.processing.system_info_collectors.environment import (
-    process_environment_telemetry,
 )
 from monkey_island.cc.services.telemetry.processing.system_info_collectors.hostname import (
     process_hostname_telemetry,
@@ -24,7 +20,6 @@ logger = logging.getLogger(__name__)
 
 SYSTEM_INFO_COLLECTOR_TO_TELEMETRY_PROCESSORS = {
     AWS_COLLECTOR: [process_aws_telemetry],
-    ENVIRONMENT_COLLECTOR: [process_environment_telemetry],
     HOSTNAME_COLLECTOR: [process_hostname_telemetry],
     PROCESS_LIST_COLLECTOR: [check_antivirus_existence],
 }