+++ monkey/monkey_island/cc/ui/src/components/pages/RunMonkeyPage/RunOnAWS/AWSRunOptions.js
@@ -81,7 +81,7 @@ const getContents = (props) => {
         <p className="alert alert-info">
           <FontAwesomeIcon icon={faInfoCircle} style={{'marginRight': '5px'}}/>
           Not sure what this is? Not seeing your AWS EC2 instances? <a
-          href="https://github.com/guardicore/monkey/wiki/Monkey-Island:-Running-the-monkey-on-AWS-EC2-instances"
+          href="https://www.guardicore.com/infectionmonkey/docs/usage/integrations/aws-run-on-ec2-machine/"
           rel="noopener noreferrer" target="_blank">Read the documentation</a>!
         </p>
       </div>