+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 - Remove serialization of config. #1537
 - Checkbox that gave the option to not try to first move the dropper file. #1537
 - Custom singleton mutex name config option. #1589
+- Removed environment system info collector #1535
 - Azure credential collector, because it was broken (not gathering credentials). #1535
 - Custom monkey directory name config option. #1537